requirements.txt
youtube-transcript-api
google-generativeai
python-docx
python-dotenv
gradio